Even more work on Unix support for CPU-specific optimizations.
